backend/requirements.txt
fastapi>=0.110.0
mangum>=0.17.0
pandas>=2.2.0
numpy>=1.26.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
